+++ piechart.h
@@ -10,7 +10,7 @@ class PieChart : public Chart {
 public:
     explicit PieChart(QWidget *parent = nullptr) : Chart(parent) {}
 
-    virtual void drawChart(const QMap<QString, qint64> &data) override;
+    virtual void drawChart(const QMap<QString, qreal> &data) override;
     virtual void drawLegend() override;
 };
